// frontend/src/generated/Job.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Data } from "./Data";
import type { Platform } from "./Platform";

export type Job = { id: number | null, platform: Platform, external_id: string, title: string, description: string | null, url: string, budget: string | null, tags: Array<string>, raw: Data, created_at: string, updated_at: string, note: string | null, liked: boolean | null, applied_at: string | null, };