// frontend/src/generated/ListQuery.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Platform } from "./Platform";
import type { Rating } from "./Rating";
import type { Sort } from "./Sort";

export type ListQuery = { platform: Platform | null, rating: Rating | null, sort_by: Sort, page: number, page_size: number, };